backend/requirements.txt
fastmcp>=0.3.0
pydantic>=2.0.0
anthropic>=0.25.0
openai>=1.0.0
langchain>=0.1.0
mcp>=1.0.0
uvloop>=0.19.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.0.0
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
soundfile>=0.12.0
librosa>=0.10.0
numba>=0.57.0
sounddevice>=0.4.6
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.0.0
demucs>=4.0.0
jsonschema>=4.0.0
